rag_langchain/requirements.txt
streamlit
Pillow
langchain==0.0.201
pydantic  
google-cloud-aiplatform